# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ArtixPOS Admin Panel API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginRequest {
  password: string;
}

export interface LoginResponse {
  token: string;
  message: string;
}

export interface AdminMeResponse {
  admin: boolean;
  iat?: number;
}

export interface ErrorResponse {
  error: string;
}

export interface ActionResponse {
  success: boolean;
  message: string;
}

export interface DashboardMetrics {
  totalUsers: number;
  totalRevenue: number;
  newSignupsThisWeek: number;
  activeStores: number;
  totalSales: number;
  totalExpenses: number;
  bannedUsers: number;
  revenueThisMonth: number;
}

export interface AdminUser {
  id: number;
  name: string;
  email: string;
  role: string;
  tenantId?: string | null;
  isBanned: boolean | null;
  createdAt?: string | null;
  storeName?: string | null;
  businessType?: string | null;
  revenueTotal: number;
  lastActive?: string | null;
}

export interface RevenueDataPoint {
  date: string;
  revenue: number;
  sales: number;
}

export interface RevenueAnalytics {
  data: RevenueDataPoint[];
  totalRevenue: number;
  totalSales: number;
  averageOrderValue: number;
}

export interface StoreRevenueSummary {
  userId: number;
  storeName: string;
  revenue: number;
  salesCount: number;
  businessType?: string | null;
}

export interface StoreSummary {
  userId: number;
  storeName: string;
  ownerName: string;
  ownerEmail: string;
  businessType?: string | null;
  currency?: string | null;
  revenue: number;
  salesCount: number;
  productCount: number;
  isBanned?: boolean | null;
}

export interface StoreProduct {
  id: number;
  name: string;
  price?: number | null;
  category?: string | null;
  stock?: number | null;
}

export interface StoreSale {
  id: number;
  total?: number | null;
  paymentMethod?: string | null;
  createdAt?: string | null;
  itemCount: number;
}

export interface StoreExpense {
  id: number;
  description?: string | null;
  amount?: number | null;
  createdAt?: string | null;
}

export interface StoreDetail {
  store: StoreSummary;
  products: StoreProduct[];
  sales: StoreSale[];
  expenses: StoreExpense[];
  aiMemoryCount: number;
}

export interface AiUsageEntry {
  userId: number;
  tenantId?: string | null;
  storeName: string;
  memoryCount: number;
  lastActivity?: string | null;
}

export type GetUsersParams = {
  search?: string;
  role?: string;
  banned?: boolean;
};

export type GetRevenueParams = {
  period?: GetRevenuePeriod;
};

export type GetRevenuePeriod =
  (typeof GetRevenuePeriod)[keyof typeof GetRevenuePeriod];

export const GetRevenuePeriod = {
  "7d": "7d",
  "30d": "30d",
  "90d": "90d",
  "1y": "1y",
} as const;