# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ArtixPOS Admin Panel API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Admin authentication
  - name: dashboard
    description: Dashboard metrics
  - name: users
    description: User management
  - name: revenue
    description: Revenue analytics
  - name: stores
    description: Store explorer
  - name: ai
    description: AI usage stats
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /admin/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/me:
    get:
      operationId: adminMe
      tags: [auth]
      summary: Verify admin token
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Token valid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminMeResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/dashboard:
    get:
      operationId: getDashboard
      tags: [dashboard]
      summary: Get dashboard metrics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard metrics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardMetrics"

  /admin/users:
    get:
      operationId: getUsers
      tags: [users]
      summary: Get all users
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: role
          in: query
          schema:
            type: string
        - name: banned
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"

  /admin/users/{userId}/ban:
    post:
      operationId: banUser
      tags: [users]
      summary: Ban a user
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User banned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResponse"

  /admin/users/{userId}/unban:
    post:
      operationId: unbanUser
      tags: [users]
      summary: Unban a user
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User unbanned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResponse"

  /admin/revenue:
    get:
      operationId: getRevenue
      tags: [revenue]
      summary: Get revenue analytics over time
      security:
        - bearerAuth: []
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [7d, 30d, 90d, 1y]
            default: 30d
      responses:
        "200":
          description: Revenue analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevenueAnalytics"

  /admin/revenue/top-stores:
    get:
      operationId: getTopStores
      tags: [revenue]
      summary: Get top 10 stores by revenue
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Top stores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StoreRevenueSummary"

  /admin/stores:
    get:
      operationId: getStores
      tags: [stores]
      summary: Get all stores with summary
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of stores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StoreSummary"

  /admin/stores/{userId}:
    get:
      operationId: getStoreDetail
      tags: [stores]
      summary: Get detailed store info
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Store detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreDetail"

  /admin/ai-usage:
    get:
      operationId: getAiUsage
      tags: [ai]
      summary: Get AI memory usage per store
      security:
        - bearerAuth: []
      responses:
        "200":
          description: AI usage data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AiUsageEntry"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    LoginRequest:
      type: object
      properties:
        password:
          type: string
      required:
        - password

    LoginResponse:
      type: object
      properties:
        token:
          type: string
        message:
          type: string
      required:
        - token
        - message

    AdminMeResponse:
      type: object
      properties:
        admin:
          type: boolean
        iat:
          type: number
      required:
        - admin

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    ActionResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message

    DashboardMetrics:
      type: object
      properties:
        totalUsers:
          type: integer
        totalRevenue:
          type: number
        newSignupsThisWeek:
          type: integer
        activeStores:
          type: integer
        totalSales:
          type: integer
        totalExpenses:
          type: number
        bannedUsers:
          type: integer
        revenueThisMonth:
          type: number
      required:
        - totalUsers
        - totalRevenue
        - newSignupsThisWeek
        - activeStores
        - totalSales
        - totalExpenses
        - bannedUsers
        - revenueThisMonth

    AdminUser:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        tenantId:
          type: string
          nullable: true
        isBanned:
          type: boolean
          nullable: true
        createdAt:
          type: string
          nullable: true
        storeName:
          type: string
          nullable: true
        businessType:
          type: string
          nullable: true
        revenueTotal:
          type: number
        lastActive:
          type: string
          nullable: true
      required:
        - id
        - name
        - email
        - role
        - isBanned
        - revenueTotal

    RevenueDataPoint:
      type: object
      properties:
        date:
          type: string
        revenue:
          type: number
        sales:
          type: integer
      required:
        - date
        - revenue
        - sales

    RevenueAnalytics:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/RevenueDataPoint"
        totalRevenue:
          type: number
        totalSales:
          type: integer
        averageOrderValue:
          type: number
      required:
        - data
        - totalRevenue
        - totalSales
        - averageOrderValue

    StoreRevenueSummary:
      type: object
      properties:
        userId:
          type: integer
        storeName:
          type: string
        revenue:
          type: number
        salesCount:
          type: integer
        businessType:
          type: string
          nullable: true
      required:
        - userId
        - storeName
        - revenue
        - salesCount

    StoreSummary:
      type: object
      properties:
        userId:
          type: integer
        storeName:
          type: string
        ownerName:
          type: string
        ownerEmail:
          type: string
        businessType:
          type: string
          nullable: true
        currency:
          type: string
          nullable: true
        revenue:
          type: number
        salesCount:
          type: integer
        productCount:
          type: integer
        isBanned:
          type: boolean
          nullable: true
      required:
        - userId
        - storeName
        - ownerName
        - ownerEmail
        - revenue
        - salesCount
        - productCount

    StoreDetail:
      type: object
      properties:
        store:
          $ref: "#/components/schemas/StoreSummary"
        products:
          type: array
          items:
            $ref: "#/components/schemas/StoreProduct"
        sales:
          type: array
          items:
            $ref: "#/components/schemas/StoreSale"
        expenses:
          type: array
          items:
            $ref: "#/components/schemas/StoreExpense"
        aiMemoryCount:
          type: integer
      required:
        - store
        - products
        - sales
        - expenses
        - aiMemoryCount

    StoreProduct:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        price:
          type: number
          nullable: true
        category:
          type: string
          nullable: true
        stock:
          type: integer
          nullable: true
      required:
        - id
        - name

    StoreSale:
      type: object
      properties:
        id:
          type: integer
        total:
          type: number
          nullable: true
        paymentMethod:
          type: string
          nullable: true
        createdAt:
          type: string
          nullable: true
        itemCount:
          type: integer
      required:
        - id
        - itemCount

    StoreExpense:
      type: object
      properties:
        id:
          type: integer
        description:
          type: string
          nullable: true
        amount:
          type: number
          nullable: true
        createdAt:
          type: string
          nullable: true
      required:
        - id

    AiUsageEntry:
      type: object
      properties:
        userId:
          type: integer
        tenantId:
          type: string
          nullable: true
        storeName:
          type: string
        memoryCount:
          type: integer
        lastActivity:
          type: string
          nullable: true
      required:
        - userId
        - storeName
        - memoryCount