# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ArtixPOS Admin Panel API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Admin login
 */
export const AdminLoginBody = zod.object({
  password: zod.string(),
});

export const AdminLoginResponse = zod.object({
  token: zod.string(),
  message: zod.string(),
});

/**
 * @summary Verify admin token
 */
export const AdminMeResponse = zod.object({
  admin: zod.boolean(),
  iat: zod.number().optional(),
});

/**
 * @summary Get dashboard metrics
 */
export const GetDashboardResponse = zod.object({
  totalUsers: zod.number(),
  totalRevenue: zod.number(),
  newSignupsThisWeek: zod.number(),
  activeStores: zod.number(),
  totalSales: zod.number(),
  totalExpenses: zod.number(),
  bannedUsers: zod.number(),
  revenueThisMonth: zod.number(),
});

/**
 * @summary Get all users
 */
export const GetUsersQueryParams = zod.object({
  search: zod.coerce.string().optional(),
  role: zod.coerce.string().optional(),
  banned: zod.coerce.boolean().optional(),
});

export const GetUsersResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  email: zod.string(),
  role: zod.string(),
  tenantId: zod.string().nullish(),
  isBanned: zod.boolean().nullable(),
  createdAt: zod.string().nullish(),
  storeName: zod.string().nullish(),
  businessType: zod.string().nullish(),
  revenueTotal: zod.number(),
  lastActive: zod.string().nullish(),
});
export const GetUsersResponse = zod.array(GetUsersResponseItem);

/**
 * @summary Ban a user
 */
export const BanUserParams = zod.object({
  userId: zod.coerce.number(),
});

export const BanUserResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
});

/**
 * @summary Unban a user
 */
export const UnbanUserParams = zod.object({
  userId: zod.coerce.number(),
});

export const UnbanUserResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
});

/**
 * @summary Get revenue analytics over time
 */
export const getRevenueQueryPeriodDefault = `30d`;

export const GetRevenueQueryParams = zod.object({
  period: zod
    .enum(["7d", "30d", "90d", "1y"])
    .default(getRevenueQueryPeriodDefault),
});

export const GetRevenueResponse = zod.object({
  data: zod.array(
    zod.object({
      date: zod.string(),
      revenue: zod.number(),
      sales: zod.number(),
    }),
  ),
  totalRevenue: zod.number(),
  totalSales: zod.number(),
  averageOrderValue: zod.number(),
});

/**
 * @summary Get top 10 stores by revenue
 */
export const GetTopStoresResponseItem = zod.object({
  userId: zod.number(),
  storeName: zod.string(),
  revenue: zod.number(),
  salesCount: zod.number(),
  businessType: zod.string().nullish(),
});
export const GetTopStoresResponse = zod.array(GetTopStoresResponseItem);

/**
 * @summary Get all stores with summary
 */
export const GetStoresResponseItem = zod.object({
  userId: zod.number(),
  storeName: zod.string(),
  ownerName: zod.string(),
  ownerEmail: zod.string(),
  businessType: zod.string().nullish(),
  currency: zod.string().nullish(),
  revenue: zod.number(),
  salesCount: zod.number(),
  productCount: zod.number(),
  isBanned: zod.boolean().nullish(),
});
export const GetStoresResponse = zod.array(GetStoresResponseItem);

/**
 * @summary Get detailed store info
 */
export const GetStoreDetailParams = zod.object({
  userId: zod.coerce.number(),
});

export const GetStoreDetailResponse = zod.object({
  store: zod.object({
    userId: zod.number(),
    storeName: zod.string(),
    ownerName: zod.string(),
    ownerEmail: zod.string(),
    businessType: zod.string().nullish(),
    currency: zod.string().nullish(),
    revenue: zod.number(),
    salesCount: zod.number(),
    productCount: zod.number(),
    isBanned: zod.boolean().nullish(),
  }),
  products: zod.array(
    zod.object({
      id: zod.number(),
      name: zod.string(),
      price: zod.number().nullish(),
      category: zod.string().nullish(),
      stock: zod.number().nullish(),
    }),
  ),
  sales: zod.array(
    zod.object({
      id: zod.number(),
      total: zod.number().nullish(),
      paymentMethod: zod.string().nullish(),
      createdAt: zod.string().nullish(),
      itemCount: zod.number(),
    }),
  ),
  expenses: zod.array(
    zod.object({
      id: zod.number(),
      description: zod.string().nullish(),
      amount: zod.number().nullish(),
      createdAt: zod.string().nullish(),
    }),
  ),
  aiMemoryCount: zod.number(),
});

/**
 * @summary Get AI memory usage per store
 */
export const GetAiUsageResponseItem = zod.object({
  userId: zod.number(),
  tenantId: zod.string().nullish(),
  storeName: zod.string(),
  memoryCount: zod.number(),
  lastActivity: zod.string().nullish(),
});
export const GetAiUsageResponse = zod.array(GetAiUsageResponseItem);