# lib/api-zod/src/generated/types/adminMeResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ArtixPOS Admin Panel API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminMeResponse {
  admin: boolean;
  iat?: number;
}