# lib/api-zod/src/generated/types/adminUser.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ArtixPOS Admin Panel API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminUser {
  id: number;
  name: string;
  email: string;
  role: string;
  tenantId?: string | null;
  isBanned: boolean | null;
  createdAt?: string | null;
  storeName?: string | null;
  businessType?: string | null;
  revenueTotal: number;
  lastActive?: string | null;
}